pyproject.toml
[project]
name = "ml-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diskcache>=5.6.3",
    "google-genai>=1.43.0",
    "ollama>=0.6.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
